# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pet Hotel & Grooming System API",
    "description": "Contract-first OpenAPI for the Pet Hotel & Grooming System.\n\n**S1 status:** endpoints are stubs (HTTP 501) with complete request/response schemas — this document is the single source of truth for backend + frontend. Business logic, state machines and real payment/notification land in S2+.\n\nAuthoritative data model: `diagrams/puml/class_diagram_plantuml.puml`. Internal modules (payment, notification, audit) are not exposed as endpoints.",
    "version": "0.1.0-S1"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register an owner account (FR-01.1)",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in, obtain JWT (FR-01.1)",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request a password-reset email (FR-01.1)",
        "operationId": "request_password_reset_api_v1_auth_password_reset_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirm password reset with token (FR-01.1)",
        "operationId": "confirm_password_reset_api_v1_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get my profile (FR-01.4)",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Edit my name / phone (FR-01.4)",
        "operationId": "update_me_api_v1_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/accounts": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List accounts — admin (FR-01.3)",
        "operationId": "list_accounts_api_v1_auth_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountOut"
                  },
                  "type": "array",
                  "title": "Response List Accounts Api V1 Auth Accounts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/accounts/{account_id}/ban": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Ban an account — admin (FR-01.3)",
        "operationId": "ban_account_api_v1_auth_accounts__account_id__ban_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanAccountIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts/{account_id}/unban": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Unban an account — admin (FR-01.3)",
        "operationId": "unban_account_api_v1_auth_accounts__account_id__unban_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/staff": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a staff account — admin (FR-01.2)",
        "operationId": "create_staff_api_v1_auth_staff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/roles": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List roles (FR-01.2)",
        "operationId": "list_roles_api_v1_auth_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleOut"
                  },
                  "type": "array",
                  "title": "Response List Roles Api V1 Auth Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/permissions": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List permissions (FR-01.2)",
        "operationId": "list_permissions_api_v1_auth_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PermissionOut"
                  },
                  "type": "array",
                  "title": "Response List Permissions Api V1 Auth Permissions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/accounts/{account_id}/roles": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Assign a role to an account (FR-01.2)",
        "operationId": "assign_role_api_v1_auth_accounts__account_id__roles_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts/{account_id}/roles/{role_id}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke a role from an account (FR-01.2)",
        "operationId": "revoke_role_api_v1_auth_accounts__account_id__roles__role_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reports/abnormal-cancellations": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Abnormal cancellation report — admin (FR-01.3)",
        "operationId": "abnormal_cancellation_report_api_v1_auth_reports_abnormal_cancellations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationReportOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "displayName": {
            "type": "string",
            "title": "Displayname"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "displayName",
          "status",
          "createdAt"
        ],
        "title": "AccountOut"
      },
      "AccountStatus": {
        "type": "string",
        "enum": [
          "Active",
          "Banned",
          "Disabled"
        ],
        "title": "AccountStatus"
      },
      "AssignRoleIn": {
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid",
            "title": "Roleid"
          }
        },
        "type": "object",
        "required": [
          "roleId"
        ],
        "title": "AssignRoleIn"
      },
...
      "BanAccountIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "examples": [
              "異常取消過多"
            ]
          }
        },
        "type": "object",
        "title": "BanAccountIn"
      },
...
      "CancellationReportOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CancellationReportRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "rows"
        ],
        "title": "CancellationReportOut"
      },
      "CancellationReportRow": {
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid",
            "title": "Bookingid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "title": "Ownerid"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "title": "Cancelledat"
          },
          "cancelReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelreason"
          },
          "refunded": {
            "type": "boolean",
            "title": "Refunded"
          }
        },
        "type": "object",
        "required": [
          "bookingId",
          "ownerId",
          "cancelledAt",
          "refunded"
        ],
        "title": "CancellationReportRow"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "P@ssw0rd!"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn"
      },
...
      "MessageOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "ok"
            ]
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageOut",
        "description": "Generic acknowledgement payload."
      },
...
      "PasswordResetConfirmIn": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "token",
          "newPassword"
        ],
        "title": "PasswordResetConfirmIn"
      },
      "PasswordResetRequestIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "PasswordResetRequestIn"
      },
...
      "PermissionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code"
        ],
        "title": "PermissionOut"
      },
...
      "ProfileUpdateIn": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "title": "ProfileUpdateIn",
        "description": "FR-01.4 — users may edit their own name + phone."
      },
...
      "RegisterIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "examples": [
              "P@ssw0rd!"
            ]
          },
          "displayName": {
            "type": "string",
            "title": "Displayname",
            "examples": [
              "王小明"
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "examples": [
              "0912345678"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "title": "RegisterIn"
      },
...
      "RoleAssignmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "accountId": {
            "type": "string",
            "format": "uuid",
            "title": "Accountid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid",
            "title": "Roleid"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Assignedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "accountId",
          "roleId",
          "assignedAt"
        ],
        "title": "RoleAssignmentOut"
      },
      "RoleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "RoleOut"
      },
...
      "StaffCreateIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "displayName": {
            "type": "string",
            "title": "Displayname"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "roleName": {
            "type": "string",
            "title": "Rolename",
            "description": "FrontDesk | Groomer | Admin",
            "examples": [
              "FrontDesk"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "displayName",
          "roleName"
        ],
        "title": "StaffCreateIn",
        "description": "FR-01.2 — admin creates a staff account with a role."
      },
...
      "TokenOut": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "examples": [
              "<jwt>"
            ]
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin",
            "examples": [
              3600
            ]
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "title": "TokenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer JWT (enforced in S2)",
        "scheme": "bearer"
      }
    }
  },
...
}